requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
cssselect2==0.7.0
Django==5.1.1
django-localflavor==4.0
django-parler==2.3
django-rosetta==0.10.1
html5lib==1.1
idna==3.8
lxml==5.3.0
oscrypto==1.3.0
pillow==10.4.0
polib==1.2.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.0.0
pypng==0.20220715.0
python-bidi==0.6.0
python-stdnum==1.20
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==10.10.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.16
